fix(workflow): ignore markers in code blocks
